Add final keyword to main() arguments.
It's never a good idea to reassign a function's arguments (since this
cannot have any effect outside the function, but that fact isn't obvious
from reading the code). Adding the final keyword makes attempting to do
so a compile-time error instead of a subtle runtime bug. Because of
this, I have Eclipse set up to automatically add the keyword to every
function's arguments whenever I save.